[progress] add body-parser to signup route
