+++ server/models/user.js
@@ -18,45 +18,49 @@ var UserSchema = new mongoose.Schema({
   salt: String
 });
 
-UserSchema.methods.comparePasswords = function (candidatePassword) {
-  var defer = Q.defer();
-  var savedPassword = this.password;
-  bcrypt.compare(candidatePassword, savedPassword, function (err, isMatch) {
-    if (err) {
-      defer.reject(err);
-    } else {
-      defer.resolve(isMatch);
-    }
-  });
-  return defer.promise;
-};
+// UserSchema.methods.comparePasswords = function (candidatePassword) {
+//   var defer = Q.defer();
+//   var savedPassword = this.password;
+//   bcrypt.compare(candidatePassword, savedPassword, function (err, isMatch) {
+//     if (err) {
+//       defer.reject(err);
+//     } else {
+//       defer.resolve(isMatch);
+//     }
+//   });
+//   return defer.promise;
+// };
 
-UserSchema.pre('save', function (next) {
-  var user = this;
+// UserSchema.pre('save', function (next) {
+//   var user = this;
 
-  // only hash the password if it has been modified (or is new)
-  if (!user.isModified('password')) {
-    return next();
-  }
+//   // only hash the password if it has been modified (or is new)
+//   if (!user.isModified('password')) {
+//     return next();
+//   }
 
-  // generate a salt
-  bcrypt.genSalt(SALT_WORK_FACTOR, function(err, salt) {
-    if (err) {
-      return next(err);
-    }
+//   // generate a salt
+//   bcrypt.genSalt(SALT_WORK_FACTOR, function(err, salt) {
+//     if (err) {
+//       return next(err);
+//     }
 
-    // hash the password along with our new salt
-    bcrypt.hash(user.password, salt, null, function(err, hash) {
-      if (err) {
-        return next(err);
-      }
+//     // hash the password along with our new salt
+//     bcrypt.hash(user.password, salt, null, function(err, hash) {
+//       if (err) {
+//         return next(err);
+//       }
 
-      // override the cleartext password with the hashed one
-      user.password = hash;
-      user.salt = salt;
-      next();
-    });
-  });
-});
+//       // override the cleartext password with the hashed one
+//       user.password = hash;
+//       user.salt = salt;
+//       next();
+//     });
+//   });
+// });
 
-module.exports = mongoose.model('users', UserSchema);
+// create model class
+var ModelClass = mongoose.model('users', UserSchema);
+
+//export model
+module.exports = ModelClass;
