+++ server/server.js
@@ -25,7 +25,7 @@ db.once('open', function() {
 //App
 app.use(morgan('combined'));
 app.use(express.static(path.join(__dirname, '../client')));
-//app.use(bodyParser.json({type: '*/*'}));
+// app.use(bodyParser.json({type: 'application/*+json'}));
 router(app);
 
 //Server
